+++ bin/just-import-git.py
@@ -230,8 +230,14 @@ def handle_import(args):
         foreign_config_file = args.foreign_repository_config
     else:
         foreign_config_file = get_repository_config_file(srcdir)
-    with open(foreign_config_file) as f:
-        foreign_config = json.load(f)
+    if args.plain:
+        foreign_config = { "main": "",
+                           "repositories": {"": {"repository":
+                                                 {"type": "file",
+                                                  "path": "." }}}}
+    else:
+        with open(foreign_config_file) as f:
+            foreign_config = json.load(f)
     foreign_repos = foreign_config.get("repositories", {})
     if args.foreign_repository_name:
         foreign_name = args.foreign_repository_name
@@ -292,6 +298,13 @@ def main():
         dest="foreign_repository_config",
         help="Repository-description file in the repository to import",
         metavar="relative-path"
+    )
+    parser.add_argument(
+        "--plain",
+        action="store_true",
+        help=
+        "Pretend the remote repository description is the canonical"
+        + " single-repository one",
     )
     parser.add_argument(
         "--as",